fixed issue with signed and unsigned comparisons in templated boundary functions not working as intended. Fixed issue with concentration boundary, now it works on grid cells only. Fixed it so that the program now gives feedback if a random particle position is requested, but no positions are allowed. Major updates to felix apps

+++ src/soskmc/averageheightboundary.cpp
@@ -9,7 +9,7 @@ AverageHeightBoundary::AverageHeightBoundary(SOSSolver &solver,
                                              const uint span,
                                              const uint yspan,
                                              Boundary::orientations orientation,
-                                             const double location) :
+                                             const uint location) :
     SOSBoundary(solver, orientation),
     m_averageHeightDepth(averageHeightDepth == 0 ? span : averageHeightDepth),
     m_dim(dim),
